Conflicts:
	bgpd/bgp_route.c

+++ lib/spf_backoff.c
@@ -226,8 +226,9 @@ void spf_backoff_show(struct spf_backoff *backoff, struct vty *vty,
 	if (backoff->t_holddown) {
 		struct timeval remain =
 			thread_timer_remain(backoff->t_holddown);
-		vty_out(vty, "%s                   Still runs for %ld msec\n",
-			prefix, remain.tv_sec * 1000 + remain.tv_usec / 1000);
+		vty_out(vty, "%s                   Still runs for %lld msec\n",
+			prefix, (long long)remain.tv_sec * 1000
+				+ remain.tv_usec / 1000);
 	} else {
 		vty_out(vty, "%s                   Inactive\n", prefix);
 	}
@@ -237,8 +238,9 @@ void spf_backoff_show(struct spf_backoff *backoff, struct vty *vty,
 	if (backoff->t_timetolearn) {
 		struct timeval remain =
 			thread_timer_remain(backoff->t_timetolearn);
-		vty_out(vty, "%s                   Still runs for %ld msec\n",
-			prefix, remain.tv_sec * 1000 + remain.tv_usec / 1000);
+		vty_out(vty, "%s                   Still runs for %lld msec\n",
+			prefix, (long long)remain.tv_sec * 1000
+				+ remain.tv_usec / 1000);
 	} else {
 		vty_out(vty, "%s                   Inactive\n", prefix);
 	}
